Fixed signed/unsigned warnings

+++ library/blob.c
@@ -391,7 +391,9 @@ int eblob_check_record(const struct eblob_base_ctl *bctl,
 	assert(bctl != NULL);
 	assert(bctl->back != NULL);
 
-	const unsigned long long bctl_size = bctl->data_size > bctl->data_offset ? bctl->data_size : bctl->data_offset;
+	const unsigned long long bctl_size = bctl->data_size > (unsigned long long)bctl->data_offset ?
+		bctl->data_size :
+		(unsigned long long)bctl->data_offset;
 
 	/*
 	 * Check record itself
